+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.0] - 2022-12-06
+
+### Changed
+
+- Improve debugging information in `SavedModelPolicy`.
+- Resample bots at the beginning of scenario episodes.
+
+### Fixed
+
+- Initialize `Transform` before any other component.
+  [#84](https://github.com/deepmind/meltingpot/issues/84)
+  [#24](https://github.com/deepmind/meltingpot/issues/24)
+
+### Added
+
+- New substrate "Hidden Agenda" and its scenarios.
+
 
 ## [2.0.0] - 2022-11-25